Fixed many errors

+++ Your_Location_In_The_Nation/src/FindLocation/CostOfLivingComparator.java
@@ -6,6 +6,6 @@ public class CostOfLivingComparator implements Comparator<Location> {
 	@Override
 	//descending order
 	public int compare(Location L1, Location L2) {
-		return L2.getCostOfLivingRent() - L1.getCostOfLivingRent();
+		return (int)(L2.getCostOfLivingRent() - L1.getCostOfLivingRent());
 	}
 }
