+++ Your_Location_In_The_Nation/src/FindLocation/CostOfLivingRentComparator.java
@@ -5,7 +5,7 @@ import LocationModel.Location;
 public class CostOfLivingRentComparator implements Comparator<Location> {
 	@Override
 	//descending order
-	public int compare(Location L1, Location L2) {
+	public float compare(Location L1, Location L2) {
 		return L2.getCostOfLivingRent() - L1.getCostOfLivingRent();
 	}
 }
